type tag narrowing
typecheck refactor and cleanup
