+++ baml_language/crates/baml_lsp_actions_tests/test_files/syntax/headers/ai_content_pipeline.baml
@@ -98,17 +98,6 @@ function AIContentPipeline(video_id: string, zoom_meeting_id: string, transcript
 //     │ 
 //     │ Note: Error code: E0003
 // ────╯
-// Error: Expected `bool`, found `string?`
-//     ╭─[ headers_ai_content_pipeline.baml:17:18 ]
-//     │
-//  17 │ ╭─▶     let summary = if (transcript) {
-//     ┆ ┆   
-//  33 │ ├─▶     };
-//     │ │            
-//     │ ╰──────────── Expected `bool`, found `string?`
-//     │     
-//     │     Note: Error code: E0001
-// ────╯
 // Error: Unknown variable `SummarizeVideo`
 //     ╭─[ headers_ai_content_pipeline.baml:21:17 ]
 //     │
@@ -162,17 +151,6 @@ function AIContentPipeline(video_id: string, zoom_meeting_id: string, transcript
 //     │                               ╰────── Unknown variable `SaveDraft`
 //     │ 
 //     │ Note: Error code: E0003
-// ────╯
-// Error: Expected `bool`, found `unknown | string`
-//     ╭─[ headers_ai_content_pipeline.baml:50:20 ]
-//     │
-//  50 │ ╭─▶     let pr_status = if (summary) {
-//     ┆ ┆   
-//  57 │ ├─▶     };
-//     │ │            
-//     │ ╰──────────── Expected `bool`, found `unknown | string`
-//     │     
-//     │     Note: Error code: E0001
 // ────╯
 // Error: Unknown variable `CreatePR`
 //     ╭─[ headers_ai_content_pipeline.baml:52:18 ]
